+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added ability to log cheaters on your team and the enemy team [#84](https://github.com/cheshire137/competiwatch-desktop/issues/84)
 - Added percentages to all bar charts on the Trends tab [#50](https://github.com/cheshire137/competiwatch-desktop/issues/50)
+- Removed separate "Manage seasons" page, putting the "Add a season" form on the Accounts tab
 - Fixed JavaScript errors when adding a new account on the Accounts tab [#96](https://github.com/cheshire137/competiwatch-desktop/issues/96)
 - Replaced some bar charts with stacked bar charts on the Trends tab
 - Increased default height of window to reduce scrolling when logging a match